Digit_Recogintion/requirements.txt
tensorflow>=2.12.0,<2.16.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.0.0
opencv-python>=4.5.0